Response models organised into Models folder
