+++ lib/Models/ExternalWalletResponse.dart
@@ -1,0 +1,16 @@
+import 'package:razorpay_web/razorpay_web.dart';
+
+class ExternalWalletResponse {
+  String? walletName;
+
+  ExternalWalletResponse(this.walletName);
+
+  static ExternalWalletResponse fromMap(Map<dynamic, dynamic> map) {
+    String? walletName = map["external_wallet"] as String?;
+    return ExternalWalletResponse(walletName);
+  }
+
+  Json get toJson => {
+        'external_wallet': walletName,
+      };
+}
